Move RepairCommand back into Cli.java
